+++ backend/src/main/java/com/mcmasterbaja/FileAnalyzeResource.java
@@ -5,8 +5,9 @@ import com.mcmasterbaja.analyzer.AnalyzerFactory;
 import com.mcmasterbaja.exceptions.InvalidArgumentException;
 import com.mcmasterbaja.live.Serial;
 import com.mcmasterbaja.model.AnalyzerParams;
-import com.mcmasterbaja.storage.FileMetadataService;
-import com.mcmasterbaja.storage.StorageService;
+import com.mcmasterbaja.services.FileMetadataService;
+import com.mcmasterbaja.services.StorageService;
+
 import jakarta.inject.Inject;
 import jakarta.ws.rs.BeanParam;
 import jakarta.ws.rs.GET;
