+++ backend/src/main/java/com/mcmasterbaja/FileDeleteResource.java
@@ -1,12 +1,13 @@
 package com.mcmasterbaja;
 
-import com.mcmasterbaja.storage.StorageService;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.PathParam;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.jboss.logging.Logger;
+
+import com.mcmasterbaja.services.StorageService;
 
 @jakarta.ws.rs.Path("/delete")
 public class FileDeleteResource {
