+++ backend/src/main/java/com/mcmasterbaja/FileUploadResource.java
@@ -2,7 +2,8 @@ package com.mcmasterbaja;
 
 import com.mcmasterbaja.binary_csv.BinaryToCSV;
 import com.mcmasterbaja.exceptions.StorageException;
-import com.mcmasterbaja.storage.StorageService;
+import com.mcmasterbaja.services.StorageService;
+
 import jakarta.inject.Inject;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.POST;
