+++ backend/src/main/java/com/mcmasterbaja/services/DefaultFileMetadataService.java
@@ -1,4 +1,4 @@
-package com.mcmasterbaja.storage;
+package com.mcmasterbaja.services;
 
 import com.drew.imaging.mp4.Mp4MetadataReader;
 import com.drew.metadata.Tag;
