+++ backend/src/main/java/com/mcmasterbaja/services/FileMetadataService.java
@@ -1,4 +1,4 @@
-package com.mcmasterbaja.storage;
+package com.mcmasterbaja.services;
 
 import java.nio.file.Path;
 import java.time.LocalDateTime;
