+++ backend/src/main/java/com/mcmasterbaja/services/StorageService.java
@@ -1,4 +1,4 @@
-package com.mcmasterbaja.storage;
+package com.mcmasterbaja.services;
 
 import java.io.InputStream;
 import java.nio.file.Path;
